requirements.txt
numpy
torch
transformers
fastapi
uvicorn
streamlit
pydantic>=1.10,<3
pytest
python-dotenv
# Optional Gemini client (uncomment if you want real Gemini calls)
# google-genai
